Fix memory issue when generating FSP matrices.

+++ src/cme_util.cu
@@ -4,9 +4,9 @@ namespace cuFSP{
 
     __device__ __host__
     void indx2state(size_t indx, int *state, size_t dim, size_t *fsp_bounds) {
-        for (size_t i{1}; i <= dim; i++) {
-            state[i - 1] = indx % (fsp_bounds[i - 1] + 1);
-            indx = indx / (fsp_bounds[i - 1] + 1);
+        for (int i{1}; i <= dim; i++) {
+            state[i - 1] = (int) indx % ((int) fsp_bounds[i - 1] + 1);
+            indx = indx / ((int) fsp_bounds[i - 1] + 1);
         }
     }
 
@@ -14,7 +14,7 @@ namespace cuFSP{
     int state2indx(int *state, size_t dim, size_t *fsp_bounds) {
         int indx = 0;
         int nprod = 1;
-        for (size_t i{1}; i <= dim; ++i) {
+        for (int i{1}; i <= dim; ++i) {
             indx += state[i - 1] * nprod;
             nprod *= (fsp_bounds[i - 1] + 1);
         }
